requirements.txt
pyrogram==2.0.106
pyromod==1.5
python-dotenv==1.0.0
TgCrypto==1.2.5